stop writing out mouse messages into logs
Due to the nature of the code as of now:
- its polling mouse, 
- and say something about it on stdout _and_ into logs

```
sudo grep 'Found mouse' /var/log/syslog | wc -l
3867491
```
It obviously ends with a lot of log write event, thus slowing the code and dumps a lot of garbage in logs. It is helpful, if you want debug logs of course. As of now there's no default log level vs debug log setting/parameter. I've just simply made to stop writing out mouse positions info unless autowaybar runs in a terminal (eg. tty). So if you run it with a bind with hyprland (or any other wayland compositor people using waybar at), it won't dump ¬4 million lines into the logs in a month :)

Of course, the more elegant way is have the concept of what log info should be visible on in terminal vs log at which log level. Until that time comes this helps.